requirements.txt
chardet
cocoNLP
cryptography
Flask
gevent
greenlet
gunicorn
Jinja2
markupsafe
openpyxl
pandas
phone
phonenumbers
pycryptodome
PyMySQL
requests
redis
regex
urllib3
Werkzeug
xlrd
XlsxWriter
